requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
openai==1.3.7
pypdf2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2 